renamed components folder to be more accurately titled, commented out currently non-functional post request (fix later)
