+++ client/src/components/RatingsAndReviews/functions/identifyChar.js
